refactor: MVPに書き換えた
