+++ ios-training/WeatherView/WeatherPresenterInput.swift
@@ -1,5 +1,5 @@
 //
-//  WeatherViewEventHandler.swift
+//  WeatherPresenterInput.swift
 //  ios-training
 //
 //  Created by 垣本 桃弥 on 2024/02/14.
@@ -7,10 +7,11 @@
 
 import Foundation
 
-/// `WeatherView`のイベント処理を行う
-protocol WeatherViewEventHandler: AnyObject {
+protocol WeatherPresenterInput {
     /// Close ボタンが押された時の処理
     func didTapCloseButton()
     /// Reload ボタンが押された時の処理
     func didTapReloadButton()
+    /// アラートのRetryボタンが押された時の処理
+    func didTapRetry()
 }
