+++ ios-training/WeatherView/WeatherViewController.swift
@@ -9,101 +9,215 @@ import UIKit
 
 /// 天気を表示する画面
 final class WeatherViewController: UIViewController {
-    // MARK: Properties
+    // MARK: Properties - UI
     
-    private let myView: WeatherViewProtocol
+    private let weatherFrame: UIView = {
+        let view = UIView()
+        view.backgroundColor = .clear
+        return view
+    }()
     
-    // MARK: Dependencies
+    private let weatherImageView: UIImageView = {
+        let imageView = UIImageView()
+        imageView.contentMode = .scaleAspectFit
+        return imageView
+    }()
     
-    private let weatherInfoRepository: WeatherInfoRepositoryProtocol
+    private let weatherImagePlaceholderLabel: UILabel = {
+        let label = UILabel()
+        label.textColor = .label
+        label.font = .preferredFont(forTextStyle: .subheadline)
+        label.textAlignment = .center
+        label.numberOfLines = 2
+        label.text = "Reloadを押して\n天気を読み込む"
+        return label
+    }()
+    
+    private let minimumTemperatureLabel: UILabel = {
+        let label = UILabel()
+        label.textColor = .systemBlue
+        label.font = .preferredFont(forTextStyle: .body)
+        label.textAlignment = .center
+        label.text = "--"
+        return label
+    }()
+    
+    private let highTemperatureLabel: UILabel = {
+        let label = UILabel()
+        label.textColor = .systemRed
+        label.font = .preferredFont(forTextStyle: .body)
+        label.textAlignment = .center
+        label.text = "--"
+        return label
+    }()
+    
+    private let closeButton: UIButton = {
+        let button = UIButton(type: .system)
+        button.setTitle("Close", for: .normal)
+        return button
+    }()
+    
+    private let reloadButton: UIButton = {
+        let button = UIButton(type: .system)
+        button.setTitle("Reload", for: .normal)
+        return button
+    }()
+    
+    private let activityIndicator = UIActivityIndicatorView(style: .large)
+    
+    // MARK: Properties - Dependencies
+    
+    private var presenter: WeatherPresenterInput!
     
     // MARK: Lifecycle
     
-    init(view: WeatherViewProtocol = WeatherView(), weatherInfoRepository: WeatherInfoRepositoryProtocol) {
-        self.myView = view
-        self.weatherInfoRepository = weatherInfoRepository
-        super.init(nibName: nil, bundle: nil)
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
+    func inject(presenter: WeatherPresenterInput) {
+        self.presenter = presenter
     }
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        view = myView
-        myView.eventHandler = self
-        
-        NotificationCenter.default.addObserver(
-            forName: UIApplication.willEnterForegroundNotification,
-            object: nil,
-            queue: nil
-        ) { [weak self] _ in
-            self?.loadWeather()
-        }
+        view.backgroundColor = .systemBackground
+        addSubviewUIs()
+        setUpLayout()
+        registerEvent()
     }
 }
 
-// MARK: - WeatherViewEventHandler
+// MARK: - WeatherViewPresenterInput
 
-extension WeatherViewController: WeatherViewEventHandler {
-    func didTapCloseButton() {
+extension WeatherViewController: WeatherPresenterOutput {
+    func dismiss() {
         dismiss(animated: true)
     }
     
-    func didTapReloadButton() {
-        loadWeather()
+    func 読み込み前() {
+        weatherImagePlaceholderLabel.isHidden = true
+        activityIndicator.startAnimating()
+        closeButton.isEnabled = false
+        reloadButton.isEnabled = false
+    }
+    
+    func 読み込み完了() {
+        activityIndicator.stopAnimating()
+        closeButton.isEnabled = true
+        reloadButton.isEnabled = true
+    }
+    
+    func 読み込み成功(weatherInfo: WeatherInfo) {
+        weatherImageView.image = .weatherImage(for: weatherInfo.weather)
+        weatherImageView.tintColor = self.imageTint(for: weatherInfo.weather)
+        minimumTemperatureLabel.text = weatherInfo.minimumTemperature.description
+        highTemperatureLabel.text = weatherInfo.highTemperature.description
+    }
+    func 読み込み失敗() {
+        let alert = AlertMaker.retryOrCancelAlert(
+            title: "天気の取得に失敗しました",
+            message: "再試行しますか？",
+            didTapRetry: { [unowned self] _ in
+                presenter.didTapRetry()
+            },
+            didTapCancel: nil
+        )
+        present(alert, animated: true)
+        weatherImageView.image = nil
+        weatherImagePlaceholderLabel.text = "取得エラー"
+        weatherImagePlaceholderLabel.isHidden = false
+        minimumTemperatureLabel.text = "--"
+        highTemperatureLabel.text = "--"
     }
 }
 
 // MARK: - Private
 
 private extension WeatherViewController {
-    /// 天気を読み込む
-    func loadWeather() {
-        // 読み込み前
-        myView.weatherImagePlaceholderLabel.isHidden = true
-        myView.activityIndicator.startAnimating()
-        myView.closeButton.isEnabled = false
-        myView.reloadButton.isEnabled = false
-        DispatchQueue.global().async {
-            defer {
-                // 読み込み完了
-                DispatchQueue.main.async {
-                    self.myView.activityIndicator.stopAnimating()
-                    self.myView.closeButton.isEnabled = true
-                    self.myView.reloadButton.isEnabled = true
-                }
-            }
-            do {
-                let weatherInfo = try self.weatherInfoRepository.fetch(at: "tokyo", date: Date())
-                // 読み込み成功
-                DispatchQueue.main.async {
-                    self.myView.weatherImageView.image = .weatherImage(for: weatherInfo.weather)
-                    self.myView.weatherImageView.tintColor = self.imageTint(for: weatherInfo.weather)
-                    self.myView.minimumTemperatureLabel.text = weatherInfo.minimumTemperature.description
-                    self.myView.highTemperatureLabel.text = weatherInfo.highTemperature.description
-                }
-            } catch {
-                // 読み込み失敗
-                DispatchQueue.main.async {
-                    let alert = AlertMaker.retryOrCancelAlert(
-                        title: "天気の取得に失敗しました",
-                        message: "再試行しますか？",
-                        didTapRetry: { [unowned self] _ in
-                            self.loadWeather()
-                        },
-                        didTapCancel: nil
-                    )
-                    self.present(alert, animated: true)
-                    self.myView.weatherImageView.image = nil
-                    self.myView.weatherImagePlaceholderLabel.text = "取得エラー"
-                    self.myView.weatherImagePlaceholderLabel.isHidden = false
-                    self.myView.minimumTemperatureLabel.text = "--"
-                    self.myView.highTemperatureLabel.text = "--"
-                }
-            }
-        }
+    /// UIをaddSubViewする処理
+    func addSubviewUIs() {
+        view.addSubview(weatherFrame)
+        view.addSubview(weatherImageView)
+        view.addSubview(weatherImagePlaceholderLabel)
+        view.addSubview(minimumTemperatureLabel)
+        view.addSubview(highTemperatureLabel)
+        view.addSubview(closeButton)
+        view.addSubview(reloadButton)
+        view.addSubview(activityIndicator)
+    }
+    
+    /// オートレイアウトの設定処理
+    func setUpLayout() { // swiftlint:disable:this function_body_length
+        let safeArea = view.safeAreaLayoutGuide
+        
+        weatherFrame.translatesAutoresizingMaskIntoConstraints = false
+        NSLayoutConstraint.activate([
+            weatherFrame.centerXAnchor.constraint(equalTo: safeArea.centerXAnchor),
+            weatherFrame.centerYAnchor.constraint(equalTo: safeArea.centerYAnchor),
+            weatherFrame.widthAnchor.constraint(equalTo: safeArea.widthAnchor, multiplier: 0.5)
+        ])
+        
+        weatherImageView.translatesAutoresizingMaskIntoConstraints = false
+        NSLayoutConstraint.activate([
+            weatherImageView.topAnchor.constraint(equalTo: weatherFrame.topAnchor),
+            weatherImageView.centerXAnchor.constraint(equalTo: weatherFrame.centerXAnchor),
+            weatherImageView.widthAnchor.constraint(equalTo: weatherFrame.widthAnchor),
+            weatherImageView.heightAnchor.constraint(equalTo: weatherImageView.widthAnchor, multiplier: 1.0)
+        ])
+        
+        weatherImagePlaceholderLabel.translatesAutoresizingMaskIntoConstraints = false
+        NSLayoutConstraint.activate([
+            weatherImagePlaceholderLabel.topAnchor.constraint(equalTo: weatherImageView.topAnchor),
+            weatherImagePlaceholderLabel.bottomAnchor.constraint(equalTo: weatherImageView.bottomAnchor),
+            weatherImagePlaceholderLabel.leadingAnchor.constraint(equalTo: weatherImageView.leadingAnchor),
+            weatherImagePlaceholderLabel.trailingAnchor.constraint(equalTo: weatherImageView.trailingAnchor)
+        ])
+        
+        minimumTemperatureLabel.translatesAutoresizingMaskIntoConstraints = false
+        NSLayoutConstraint.activate([
+            minimumTemperatureLabel.topAnchor.constraint(equalTo: weatherImageView.bottomAnchor),
+            minimumTemperatureLabel.bottomAnchor.constraint(equalTo: weatherFrame.bottomAnchor),
+            minimumTemperatureLabel.leadingAnchor.constraint(equalTo: weatherImageView.leadingAnchor),
+            minimumTemperatureLabel.widthAnchor.constraint(equalTo: weatherImageView.widthAnchor, multiplier: 0.5)
+        ])
+        
+        highTemperatureLabel.translatesAutoresizingMaskIntoConstraints = false
+        NSLayoutConstraint.activate([
+            highTemperatureLabel.topAnchor.constraint(equalTo: weatherImageView.bottomAnchor),
+            highTemperatureLabel.trailingAnchor.constraint(equalTo: weatherImageView.trailingAnchor),
+            highTemperatureLabel.widthAnchor.constraint(equalTo: weatherImageView.widthAnchor, multiplier: 0.5)
+        ])
+        
+        closeButton.translatesAutoresizingMaskIntoConstraints = false
+        NSLayoutConstraint.activate([
+            closeButton.topAnchor.constraint(equalTo: minimumTemperatureLabel.bottomAnchor, constant: 80),
+            closeButton.centerXAnchor.constraint(equalTo: minimumTemperatureLabel.centerXAnchor)
+        ])
+        
+        reloadButton.translatesAutoresizingMaskIntoConstraints = false
+        NSLayoutConstraint.activate([
+            reloadButton.topAnchor.constraint(equalTo: highTemperatureLabel.bottomAnchor, constant: 80),
+            reloadButton.centerXAnchor.constraint(equalTo: highTemperatureLabel.centerXAnchor)
+        ])
+        
+        activityIndicator.translatesAutoresizingMaskIntoConstraints = false
+        NSLayoutConstraint.activate([
+            activityIndicator.topAnchor.constraint(equalTo: safeArea.topAnchor),
+            activityIndicator.bottomAnchor.constraint(equalTo: safeArea.bottomAnchor),
+            activityIndicator.leadingAnchor.constraint(equalTo: safeArea.leadingAnchor),
+            activityIndicator.trailingAnchor.constraint(equalTo: safeArea.trailingAnchor)
+        ])
+    }
+    
+    /// 画面のイベント処理を登録する
+    func registerEvent() {
+        closeButton.addTarget(self, action: #selector(didTapCloseButton), for: .touchUpInside)
+        reloadButton.addTarget(self, action: #selector(didTapReloadButton), for: .touchUpInside)
+    }
+    
+    @objc private func didTapCloseButton() {
+        presenter.didTapCloseButton()
+    }
+    
+    @objc private func didTapReloadButton() {
+        presenter.didTapReloadButton()
     }
     
     /// `weatherImageView.tintColor` に指定するための色を返す
@@ -121,11 +235,11 @@ private extension WeatherViewController {
 
 // MARK: - Preview
 
-#Preview {
-    WeatherViewController(
-        weatherInfoRepository: WeatherInfoRepository(
-            apiEncoder: YumemiWeatherAPIEncoder(),
-            apiDecoder: YumemiWeatherAPIDecoder()
-        )
-    )
-}
+//#Preview {
+//    WeatherViewController(
+//        weatherInfoRepository: WeatherInfoRepository(
+//            apiEncoder: YumemiWeatherAPIEncoder(),
+//            apiDecoder: YumemiWeatherAPIDecoder()
+//        )
+//    )
+//}
